refactor(cli): extract sandbox status action
