+++ src/lib/nemoclaw-runtime-bridge.ts
@@ -6,7 +6,6 @@
 export interface NemoClawRuntimeBridge {
   sandboxDestroy: (sandboxName: string, args?: string[]) => Promise<void>;
   sandboxRebuild: (sandboxName: string, args?: string[]) => Promise<void>;
-  sandboxStatus: (sandboxName: string) => Promise<void>;
   upgradeSandboxes: (args?: string[]) => Promise<void>;
 }
